### 2017-07-21 #321 @labarba
I've read the paper, and this is not a piece of software with a research mission.

This is not eligible for publication in JOSS.
