### 2017-07-24 #321 @konstantinstadler
Hallo Lorena and Arfon,

Thanks for quick review!

The main purpose of the software is to
"make addressing research challenges significantly better (e.g., faster, easier, simpler)".
The research application is the parsing of global data-sets.

One of the challenges we face in our research (Industrial Ecology - linking environmental and economic research to address global sustainability issues) are different country classifications, or even worse, just the use of non-standardized names when sharing data (e.g. in data provided by global organizations). As an example check the different forms of North and South Korea found in various databases - all tests here are based on encountered examples: https://github.com/konstantinstadler/country_converter/blob/master/tests/test_regex_korea.txt - the other test files provide further examples of the difficulties you can face parsing global data sets. To simplify this parsing (actually to make it possible to be automated) we needed a tool to convert between different classifications and to match various forms of one country name.

We currently use the country converter in various projects, e.g. for parsing data about globally spanning shipping routes (which use non standard country names as trading partner indication), building scenarios based on economic forecasts by the International Monetary Fund (IMF provides data using country names), parsing UN data for Multi Regional Input Output Models (converting UN numeric country codes into internally used ISO3 codes) and  parsing biodiversity data given per country (again, using different country names). Because we now use the country converter for such diverse research projects, and since it has gained some interest outside our institute already (issue submitted by a person working for WIFO Vienna, some personal email request of how to use it from other research institutes), it seemed to be a good candidate for a publication in your journal.

I believe, the country converter will be of much help for researchers from various scientific fields which depend on some secondary data to be extracted from global databases. I hope, that I could make a case for this piece of software as a research tool - please let me know, if you need any further information.
