### 2017-07-24 #321 @labarba
Alright, if you could work that big-picture explanation of the research uses of the code into the paper, then we can proceed. The paper currently does not make the case.

Because JOSS publishes software articles in a wide variety of fields, it's vital for the authors to include a few sentences for a non-specialist to understand why the software is useful for research.
