### 2017-07-24 c45134d
updated JOSS description
