+++ JOSS/paper.bib
@@ -52,22 +52,6 @@ As a widely used international standard, {ISO} 3166-1 is implemented in other st
 	file = {cesr_a_936831_sm0729.docx:/home/konstans/zotbib/storage/WE733TEG/cesr_a_936831_sm0729.docx:application/vnd.openxmlformats-officedocument.wordprocessingml.document;Stadler et al. - 2014 - The ‘rest of the World’ – Estimating the Economic .pdf:/home/konstans/zotbib/storage/PW4SZKFK/Stadler et al. - 2014 - The ‘rest of the World’ – Estimating the Economic .pdf:application/pdf}
 }
 
-@article{tukker_global_2013,
-	title = {Global {Multiregional} {Input}–{Output} {Frameworks}: {An} {Introduction} and {Outlook}},
-	volume = {25},
-	issn = {0953-5314},
-	shorttitle = {Global {Multiregional} {Input}–{Output} {Frameworks}},
-	url = {http://www.tandfonline.com/doi/abs/10.1080/09535314.2012.761179},
-	doi = {10.1080/09535314.2012.761179},
-	abstract = {This review is the introduction to a special issue of Economic Systems Research on the topic of global multiregional input–output (GMRIO) tables, models, and analysis. It provides a short historical context of GMRIO development and its applications (many of which deal with environmental extensions) and presents the rationale for the major database projects presented in this special issue. Then the six papers are briefly introduced. This is followed by a concluding comparison of the characteristics of the main GMRIO databases developed thus far and an outlook of potential further developments.},
-	number = {1},
-	urldate = {2013-04-29},
-	journal = {Economic Systems Research},
-	author = {Tukker, A. and Dietzenbacher, Erik},
-	year = {2013},
-	pages = {1--19},
-	file = {Tukker_Dietzenbacher - 2013 - Global Multiregional Input–Output Frameworks.pdf:/home/konstans/zotbib/storage/QMCP6GZX/Tukker_Dietzenbacher - 2013 - Global Multiregional Input–Output Frameworks.pdf:application/pdf}
-}
 
 @article{timmer_world_2012,
 	title = {The {World} {Input}-{Output} {Database} ({WIOD})},
