+++ JOSS/paper.md
@@ -18,25 +18,36 @@ bibliography: paper.bib
 
 # Summary
 
-
-To date, several competing standards exist for country abbreviations. This
-hinders the automatic parsing of data, in particular for the construction of globally
-spanning models and databases as for example Multi Regional Input Output Models
-(@tukker_global_2013). 
+Gathering the data basis for regional to globally spanning models and scenarios
+in many scientific fields regularly requires the parsing of multiple data
+sources. In particular for data related to economic (national accounts, trade
+statistics, etc.) and environmental (climate, ecosystem descriptions, etc.)
+variables, these data are mostly categorised per country. There is, however, no
+single standard of how to name or specify individual countries. As a
+consequence, parsing routines need to be adopted for each used data source.
 
 Of the existing standards, ISO 3166-1 (see @_iso_2017) defines a two and a
 three letter code in addition to a numerical classification. The UN uses its
 own numerical classification scheme, mostly based on the ISO 3166-1 numerical
-code. To further complicate the matter, several other databases use their own country
-names to classify countries.
+code. To further complicate the matter, instead of using one of the existing
+standards, many databases use unstandardised country names to classify
+countries. For example, South Korea might be referred to as "Korea", "Korea,
+Republic of", "Republic of Korea" or "Korea, Rep." and must be distinguished
+from the various names of North Korea: "Democratic People's Republic of Korea",
+"Dem. People's Rep. of Korea", "Korea, Dem. Rep.".
 
-To ease the conversion between various forms of country names and different
-classification schemes, I developed the country converter (coco) in python 3.
-Coco converts a given country name or ISO/UN code to a specified output format.
+In a research project relying on multiple data sources, one usually decides on
+one standard name or abbreviation to be used throughout the project and than
+provide conversion routines for matching all encountered country names and
+abbreviations into the to this master classification.
+
+The country converter (coco - a python3 programme) aims to automate this step by
+providing a fully tested conversion set between the different country
+abbreviations standards and unstandardised country names.
 
 The basis of coco is a table of regular expression which match all English
 versions of country names encountered by the author. The included tests match
-these regular expression against all found names, checking for unique matching
+the regular expression against all so far encountered names, checking for unique matching
 between regular expression and country names and vice versa.
 
 The table of regular expression and linked country names and classification can
