### 2017-07-26 #321 @rgieseke
Hi, I'd be happy to be a reviewer for this.

I'm a potential user (for emissions, GDP, population and other data in Climate Change research) - and also working on subsets of similar functionality in https://github.com/openclimatedata/countrygroups and have been using and contributing to https://github.com/occrp/countrynames.
