### 2017-07-26 #321 @labarba
Don't I have to start the review first, so that the new issue gets created, and assign the reviewer there?
