### 2017-07-26 #332 @whedon
[REVIEW]: The country converter coco - a Python package for converting country names between different classification schemes
**Submitting author:** @konstantinstadler (<a href="http://orcid.org/0000-0002-1548-201X">Konstantin Stadler</a>)
**Repository:** <a href="https://github.com/konstantinstadler/country_converter" target ="_blank">https://github.com/konstantinstadler/country_converter</a>
**Version:** v0.4
