### 2017-07-26 #321 @labarba
Interesting. So `start review` does nothing if the reviewer hasn't been assigned yet. It would be nice if `whedon` would throw a message for that.
