### 2017-07-26 #332 @labarba
@rgieseke : Thanks for volunteering to review. This is the place to enter your comments, as you go through the review checklist, but you are also encouraged to open new issues in the repository of the software, if needed.
