### 2017-07-26 #321 @arfon
@labarba - I've opened a bug report here https://github.com/openjournals/whedon-api/issues/6
