### 2017-07-26 konstantinstadler/country_converter#8 @rgieseke
License question
As part of my review (https://github.com/openjournals/joss-reviews/issues/332) a question on licensing:

More than anything, the choice of license is also cultural (https://twitter.com/hadleywickham/status/873554179792355328) and it seems that permissive licenses are more popular for Python projects. Given that `country_converter` is "infrastructure" and not a scientific model, maybe it could make sense to release under BSD or MIT? This could improve its adoption, I have noted that some Python projects are reluctant to include GPL-ed code. There are of course good reasons to choose GPL and I have myself used it in scientific projects.

If you chose GPL because it in part builds on https://github.com/vincentarelbundock/countrycode, then maybe its authors should be mentioned in the [LICENSE](https://github.com/konstantinstadler/country_converter/blob/master/LICENSE) file?
