### 2017-07-26 konstantinstadler/country_converter#9 @rgieseke
Community guidelines
With regards to the "Community guidelines" question in https://github.com/openjournals/joss-reviews/issues/332

> Community guidelines: Are there clear guidelines for third parties wishing to
>1) Contribute to the software
>2) Report issues or problems with the software
>3) Seek support

Maybe adding a `CONTRIBUTING.md` file would be useful? Also mentioning the issue tracker in the `README.md` could be helpful as people might come across the project e.g. from its PyPI page.

I would also be interested to read what kind of additional groupings you would consider for inclusion.
