### 2017-07-26 konstantinstadler/country_converter#10 @rgieseke
Git version tag
>  Version: Does the release version given match the GitHub release (v0.4)?

Could you tag the version published on PyPI in Git as well? This is also helpful for the Zenodo archiving.

(https://github.com/openjournals/joss-reviews/issues/332)
