### 2017-07-27 konstantinstadler/country_converter#11 @rgieseke
How to run tests
Could you maybe add some instructions on how to run the tests? This would be helpful for people not familiar with py.test.

I used (after cloning the repo):

```
python3 -m venv venv
./venv/bin/pip install -e .
./venv/bin/pip install pytest
./venv/bin/pytest --verbose
```

(https://github.com/openjournals/joss-reviews/issues/332)
