### 2017-07-28 konstantinstadler/country_converter#13 @rgieseke
DOI for “The World Input- Output Database (WIOD).” Working Paper 
It seems the Working Paper editor added the paper to ResearchGate and assigned a doi:

https://www.researchgate.net/publication/287208987_The_World_Input-Output_Database_WIOD_Contents_Sources_and_Methods

Not sure if it's better to use this doi, adding it to the BibTeX file seems to not display the wiod.org url anymore when running pandoc.

It's the first time I notice a RG doi and there seem to have been some questions around them:

http://blog.impactstory.org/researchgate-doi/

So it's probably ok to leave the references as is, just thought I mention it.

(https://github.com/openjournals/joss-reviews/issues/332)
