### 2017-07-28 5928fc3
Typo fixes and suggestions for paper
