### 2017-07-28 56a2982
Typo fixes and suggestions
