### 2017-07-28 konstantinstadler/country_converter#15 @rgieseke
Clarify Python version in setup.py
Maybe add

```
Programming Language :: Python :: 3
```
or similar to `setup.py`

(https://github.com/openjournals/joss-reviews/issues/332)
