### 2017-07-28 konstantinstadler/country_converter#16 @rgieseke
Links to related software
Maybe add links to related software like

[pycountry](https://pypi.python.org/pypi/pycountry), [iso-3166](https://github.com/deactivated/python-iso3166) or [countrynames](https://github.com/occrp/countrynames) (for non-English names)

This could help potential users to decide which one they need and see advantages and features of `country_converter`

(https://github.com/openjournals/joss-reviews/issues/332)
