### 2017-07-28 #332 @rgieseke
The `coco` country converter is a very useful Python library for conversion of country name variants as they appear in real-world datasets that through its included command line tool and some hints for Matlab users can be used even beyond the Python eco-system. Its user base will also likely extend beyond the ones mentioned in the accompanying paper, as the functions provided by `coco` are needed in many data analysis and visualisation problems.

@konstantinstadler
Once the issues opened in the [https://github.com/konstantinstadler/country_converter/](contry_converter repo) are discussed and check boxes ticked of, this should be ready for publication.
