### 2017-07-28 konstantinstadler/country_converter#18 @rgieseke
Function naming for conversion
Another minor question that came up during testing the library. For functions like `EU28in` I was at first assuming this would be a function to check being part of a group.

Maybe something like Pandas `astype` could work as well?

https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.astype.html

`EU28as`?

But this certainly depends on familiarity and convention ...

(openjournals/joss-reviews#332)
