### 2017-08-01 e414ac4
added a contribution introduction file
