+++ country_converter/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.4'
+__version__ = '0.5dev'
