### 2017-08-01 konstantinstadler/country_converter#9 @konstantinstadler
Added the header
Communication, issues, bugs and enhancements
to README.rst and added a CONTRIBUTING.rst
