### 2017-08-01 #332 @konstantinstadler
Thank you very much for your review! I will address them in the next days in the branch joss_review
