### 2017-08-01 ea48e3d
added link to CONTRIBUTING.rst for testing in readme.rst
