### 2017-08-01 konstantinstadler/country_converter#11 @konstantinstadler
I added a section
"Running and extending the tests"
in CONTRIBUTING.rst and link to this description when I mention pytest in the README.rst
