### 2017-08-02 konstantinstadler/country_converter#18 @konstantinstadler
This makes the method names clearer. Thanks a lot for the suggestion. Renamed the coco.XXin() methods to coco.XXas()
