### 2017-08-02 695d047
changed coco.XXin() to coco.XXas()
