### 2017-08-02 konstantinstadler/country_converter#16 @konstantinstadler
Added a section in the README
