### 2017-08-02 463e8ae
restructred readme, remove jupyter checkpoints from the repo
