### 2017-08-02 2929002
added classifiers for python version and licence
