### 2017-08-02 5945d9e
addded problem description in the readme - motivation
