+++ README.rst
@@ -6,6 +6,17 @@ The country converter (coco) is a Python package to convert country names betwee
 .. image:: https://travis-ci.org/konstantinstadler/country_converter.svg?branch=master
     :target: https://travis-ci.org/konstantinstadler/country_converter
 
+Motivation             
+^^^^^^^^^^
+
+To date, there is no single standard of how to name or specify individual countries in a (meta) data description. 
+While some data sources follow ISO 3166, this standard defines a two and a three letter code in addition to a numerical classification.
+To further complicate the matter, instead of using one of the existing standards, many databases use unstandardised country names to classify countries.
+            
+The country converter (coco) automates the conversion from different standards and version of country names. 
+Internally, coco is based on a table specifying the different ISO and UN standards per country together with the official name and a regular expression which aim to match all English versions of a specific country name. 
+In addition, coco includes classification based on UN-, EU-, OECD-membership, UN regions specifications, continents and various MRIO databases (see `Classification schemes`_). 
+
 Installation
 ------------
 
@@ -291,7 +302,8 @@ You can follow me on twitter_ or mastodon_ to get the latest news about all my o
 Contributing   
 ---------------
 
-See CONTRIBUTING.rst if you want to help to improve coco.
+Want to contribute? Great! 
+Please check CONTRIBUTING.rst if you want to help to improve coco.
   
 
 Related software   
