### 2017-08-02 cfa18da
add badge for pypi and licence
