### 2017-08-02 konstantinstadler/country_converter#10 @konstantinstadler
The version uploaded an PyPI (0.4) is already tagged, I added a badge in the readme pointing to the current version available at PyPI.
