+++ README.rst
@@ -5,11 +5,12 @@ The country converter (coco) is a Python package to convert country names betwee
 
 .. image:: https://badge.fury.io/py/country_converter.svg
     :target: https://badge.fury.io/py/country_converter
+.. image:: https://zenodo.org/badge/40310849.svg
+   :target: https://zenodo.org/badge/latestdoi/40310849
 .. image:: https://img.shields.io/badge/License-GPL%20v3-blue.svg
     :target: https://www.gnu.org/licenses/gpl-3.0
 .. image:: https://travis-ci.org/konstantinstadler/country_converter.svg?branch=master
     :target: https://travis-ci.org/konstantinstadler/country_converter
-
 
 .. contents:: Table of Contents
 
