### 2017-08-02 konstantinstadler/country_converter#10 @konstantinstadler
I also added a zenodo badge for the previous major version (0.4) - a new release will follow after the JOSS review.
