### 2017-08-02 konstantinstadler/country_converter#13 @konstantinstadler
Ok. I updated the link to the wiod database - they now have their preferred citation if using wiod on the webpage.
