### 2017-08-02 konstantinstadler/country_converter#8 @konstantinstadler
I agree, that licensing is a controversial issue.
I very much agree with the GNU philosophy (https://www.gnu.org/licenses/quick-guide-gplv3.en.html) and would therefore rather keep the current licence.
