### 2017-08-02 7845238
some readme fixes
