### 2017-08-02 #332 @konstantinstadler
@rgieseke
I went through the issues and responded to all of them. Many thanks for all your comments, in particular the suggestion with the contribution file - this made me actually thinking how I want people to contribute.
So far, all changes are in the "joss_review" branch, I will merge them with the master when I get your ok.
