### 2017-08-02 konstantinstadler/country_converter#8 @rgieseke
That's great, just wanted to make sure!
The `setup.py file` lists "GPLv3 or later" in the classifier, I believe the "license" field is thus unnecessary:

> The license field is a text indicating the license covering the package where the license is not a selection from the “License” Trove classifiers. See the Classifier field.

https://docs.python.org/3/distutils/setupscript.html#additional-meta-data

Maybe the "or later" should be stated clearly also in other places?
