### 2017-08-02 ed35f73
removed licence from the classifers in setup
