### 2017-08-02 476bad4
removed licence field, included licence gplv3 in classifers
