### 2017-08-02 konstantinstadler/country_converter#8 @konstantinstadler
Ok, thanks for pointing that out. I removed the licence field and changed to just gpl3 - dont know exactly how I would handle "or later"
