### 2017-08-02 23be468
Formatting and suggestions
