+++ README.rst
@@ -16,16 +16,16 @@ The country converter (coco) is a Python package to convert country names betwee
 
 .. contents:: Table of Contents
 
-Motivation             
+Motivation
 -----------
 
-To date, there is no single standard of how to name or specify individual countries in a (meta) data description. 
+To date, there is no single standard of how to name or specify individual countries in a (meta) data description.
 While some data sources follow ISO 3166, this standard defines a two and a three letter code in addition to a numerical classification.
 To further complicate the matter, instead of using one of the existing standards, many databases use unstandardised country names to classify countries.
-            
-The country converter (coco) automates the conversion from different standards and version of country names. 
-Internally, coco is based on a table specifying the different ISO and UN standards per country together with the official name and a regular expression which aim to match all English versions of a specific country name. 
-In addition, coco includes classification based on UN-, EU-, OECD-membership, UN regions specifications, continents and various MRIO databases (see `Classification schemes`_ below). 
+
+The country converter (coco) automates the conversion from different standards and version of country names.
+Internally, coco is based on a table specifying the different ISO and UN standards per country together with the official name and a regular expression which aim to match all English versions of a specific country name.
+In addition, coco includes classification based on UN-, EU-, OECD-membership, UN regions specifications, continents and various MRIO databases (see `Classification schemes`_ below).
 
 Installation
 ------------
@@ -53,7 +53,7 @@ Usage
 Basic usage
 ^^^^^^^^^^^
 
-Use within python
+Use within Python
 """""""""""""""""
 
 Convert various country names to some standard names:
@@ -61,7 +61,9 @@ Convert various country names to some standard names:
 .. code:: python
 
     import country_converter as coco
-    some_names = ['United Rep. of Tanzania', 'DE', 'Cape Verde', '788', 'Burma', 'COG','Iran (Islamic Republic of)', 'Korea, Republic of', "Dem. People's Rep. of Korea"]
+    some_names = ['United Rep. of Tanzania', 'DE', 'Cape Verde', '788', 'Burma', 'COG',
+                  'Iran (Islamic Republic of)', 'Korea, Republic of',
+                  "Dem. People's Rep. of Korea"]
     standard_names = coco.convert(names=some_names, to='name_short')
     print(standard_names)
 
@@ -77,7 +79,9 @@ instantiating a single CountryConverter object for all conversions:
     import country_converter as coco
     cc = coco.CountryConverter()
 
-    some_names = ['United Rep. of Tanzania', 'Cape Verde', 'Burma', 'Iran (Islamic Republic of)', 'Korea, Republic of', "Dem. People's Rep. of Korea"]
+    some_names = ['United Rep. of Tanzania', 'Cape Verde', 'Burma',
+                  'Iran (Islamic Republic of)', 'Korea, Republic of',
+                  "Dem. People's Rep. of Korea"]
 
     standard_names = cc.convert(names = some_names, to = 'name_short')
     UNmembership = cc.convert(names = some_names, to = 'UNmember')
@@ -90,7 +94,7 @@ Convert between classification schemes:
 .. code:: python
 
     iso3_codes = ['USA', 'VUT', 'TKL', 'AUT', 'XXX' ]
-    iso2_codes = coco.convert(names = iso3_codes, to = 'ISO2')
+    iso2_codes = coco.convert(names=iso3_codes, to='ISO2')
     print(iso2_codes)
 
 Which results in ['US', 'VU', 'TK', 'AT', 'not found']
[…]
