### 2017-08-02 9e7b0fb
Add direct link to CONTRIBUTING.rst
