+++ CONTRIBUTING.rst
@@ -26,18 +26,18 @@ If you come across a country name which currently is not matched by the regular 
 For a good introduction in the used regular expression syntax see https://docs.python.org/3.5/library/re.html .
 In particular, make use of the or symbol "|" to include multiple regular expressions for one country.
 See the entry for "Czech Republic" for a relatively simple example, "Republic of the Congo" for a more advanced case.
-IMPORTANT: new regular expressions must not break the already present matchings. 
+IMPORTANT: new regular expressions must not break the already present matchings.
 Run the tests before issuing a pull request and consider including a test case for the new match (see below).
 
 New country classification
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-If you think a certain country classification is missing from coco, you can simply add them as new columns in country_converter/country_data.tsv. 
-For fixed country classifications (as for example the "continent") just add a new column with the corresponding name. 
+If you think a certain country classification is missing from coco, you can simply add them as new columns in country_converter/country_data.tsv.
+For fixed country classifications (as for example the "continent") just add a new column with the corresponding name.
 For a classification which changes over time (as for example "OECD") make an new column and provide the year at with the country obtained its membership.
-Optional: you can add a new property for accessing a specific membership. 
+Optional: you can add a new property for accessing a specific membership.
 See the available properties for EU28, OECD, etc in the class CountryConverter in country_converter/country_converter.py for how to do that and continue reading at
-"Changing the code base". 
+"Changing the code base".
 New classifications must also be added to the README.rst at the section "Classification schemes"
 
 Changing the code base
@@ -45,7 +45,7 @@ Changing the code base
 
 If you plan any changes to the source code of this repo, please first discuss the change you wish to make via a filing an issue (labelled Enhancement or Bug) before making a change.
 All code contribution must be provided as pull requests connected to a filed issue.
-Use numpy style docstrings_ and follow pep8_ style guide. 
+Use numpy style docstrings_ and follow pep8_ style guide.
 The latter is a requirement to pass the tests before merging a pull request.
 Since coco is already used in research projects, please aim for keeping compatibility with previous versions.
 
@@ -55,16 +55,17 @@ Since coco is already used in research projects, please aim for keeping compatib
 Running and extending the tests
 -------------------------------
 
-Before filing a pull request, make sure your changes pass all tests. 
-Coco uses the py.test_ package with the pytest-pep8_ extension for testing. 
+Before filing a pull request, make sure your changes pass all tests.
+Coco uses the py.test_ package with the pytest-pep8_ extension for testing.
 To run the tests install these two packages and run
 
 ::
+
     py.test -v -pep8
-    
+
 in the root of your local copy of coco.
 
-The included tests verify the regular expressions against names commonly found in various databases. 
+The included tests verify the regular expressions against names commonly found in various databases.
 
 These tests check
 
@@ -72,7 +73,7 @@ These tests check
 #) Do the official name uniquely match the regular expression?
 #) Do the alternative names tested so far still uniquely match the standard names?
 
-To specify a new test set just add a tab-separated file with headers "name_short" and "name_test" and provide name (corresponding to the short name in the main classification file) and the alternative name which should be tested (one pair per row in the file). 
+To specify a new test set just add a tab-separated file with headers "name_short" and "name_test" and provide name (corresponding to the short name in the main classification file) and the alternative name which should be tested (one pair per row in the file).
 If the file name starts with "test\_regex\_" it will be automatically recognised by the test functions.
 
 .. _py.test: http://pytest.org/
