### 2017-08-02 77d7242
Mention Pandas dependency
Some users might be able to modify the data file but be unfamiliar with the
Pandas/Python eco-system.
