### 2017-08-02 8cf5204
changed pandas to Pandas in rst files
