### 2017-08-02 #332 @rgieseke
Excellent! Thanks for adressing all issues and the detailed descriptions in the CONTRIBUTING file, this should be very helpful for potential contributors!

@labarba The version number will be updated (https://github.com/konstantinstadler/country_converter/issues/10) after finalizing the review, so this might need changing in the meta-data of this issue (don't know if it's actually used in the submission process)

Another question, ist there a JOSS policy already on using Zenodo's version dois?
https://doi.org/10.5281/zenodo.838035
Likely the specific version reviewed should be used, but the version doi is of course quite useful for software.
