### 2017-08-02 #332 @rgieseke
@konstantinstadler A final (minor) thing: Can you update the meta-data in  the Zenodo archive to include the license and maybe update the title (don't know how Zenodo generates this by default). Once the JOSS paper is out you can also add a reference back to the paper. I think there is a way to have a JSON meta-data file but I can't find the documentation for this right now, but you can always edit these when logging in to Zenodo.

You might also consider adding a CITATION file or notice in the Readme on how you would like users to cite `coco` if they used in their research, see e.g. https://www.software.ac.uk/blog/2013-09-02-encouraging-citation-software-introducing-citation-files (For example asking to cite the JOSS paper.)

Otherwise, once you merged the [review](https://github.com/konstantinstadler/country_converter/tree/joss_review) branch `coco` should be ready for publication in JOSS!
