### 2017-08-02 #332 @konstantinstadler
@rgieseke Great. I updated the Zenodo description. I will do the link to JOSS and the CITATION file as soon as I get a doi for coco. I will merge with master now.
Thanks again for the review! Working with the issues tracker is so much more efficient than the traditional scientific paper review process...
