### 2017-08-02 f359597
updated changelog and version number - version 0.5
