### 2017-08-02 #332 @rgieseke
@konstantinstadler I don't see changes on https://zenodo.org/record/838036 yet - can you check again?

Thanks for publishing this as a re-usable tool, I'm sure it will be quite useful to many others!
