### 2017-08-02 #332 @konstantinstadler
Yes, I see. I can save but I get an internal server error if I want to publish the changes (I will try again later). Anyhow, the current upload to Zenodo is just for the history (v0.4). I will re-upload as soon as I can do the cross-reference to the JOSS doi.
