### 2017-08-02 #332 @arfon
> Another question, ist there a JOSS policy already on using Zenodo's version dois?
https://doi.org/10.5281/zenodo.838035

Versioned DOIs are 👍 with JOSS.
