### 2017-08-02 #332 @rgieseke
@konstantinstadler Maybe a temporary problem - I thought that depending on how you set-up the Zenodo integration you just would need to tag a new release on GitHub: https://guides.github.com/activities/citable-code/
