### 2017-08-02 #332 @arfon
> @arfon In that case, it should be ready for publication with the concept doi.

Sorry, to be clear, we would like to know the explicit version of the DOI that corresponds to the work associated with this JOSS submission before proceeding.
