### 2017-08-02 #332 @rgieseke
Makes sense to have the reviewed final version archived and referenced!

Maybe the JOSS page and/or paper could still link to the concept doi as well?
