### 2017-08-02 #332 @arfon
> Maybe the JOSS page and/or paper could still link to the concept doi as well?

You're welcome to cite it in the paper. We're not able to make two links to two different archives of the software on the JOSS page sorry.
