### 2017-08-02 #332 @konstantinstadler
Sorry for complicating the stuff with the zenodo upload and doi. I know could change the meta information in zenodo.
Would it help if I now upload v0.5 (the outcome of the JOSS review) and you set this as the concept doi?
