### 2017-08-02 #332 @rgieseke
I think there only needs to be v0.5 (as it was reviewed) tagged on GitHub and archived on Zenodo (and on PyPI) - did you do a manual upload or did you use the GitHub integration?

If you used the integration I think you just have to tag v0.5.

For the JOSS publication only the reviewed archive is relevant. (Users can then find later versions via the concept doi on Zenodo, or if you mention the concept doi in the paper or Readme.)
