### 2017-08-02 #332 @konstantinstadler
Ok. You have to tag and make a release in github. I have done that, the new zenodo doi is

10.5281/zenodo.838248
