### 2017-08-02 #332 @konstantinstadler
This version is now also available in PyPI
