### 2017-08-03 #332 @arfon
@rgieseke - many thanks for your review here and to @labarba for editing this one ✨

@konstantinstadler - your paper is now accepted in JOSS and your DOI is http://dx.doi.org/10.21105/joss.00332 ⚡️ 🚀 💥
