### 2017-08-03 #332 @konstantinstadler
@rgieseke @labarba @arfon Many thanks. That was a pleasant experience - would love to see more journals using the issue approach for reviewing!
